+++ example/lib/navigation_side.dart
@@ -1,5 +1,22 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_side_menu/flutter_side_menu.dart';
+import 'package:badges/badges.dart' as badges;
+
+void main() {
+  runApp(const MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Navigation Side',
+      home: NavigationSide(),
+    );
+  }
+}
 
 class NavigationSide extends StatelessWidget {
   const NavigationSide({Key? key}) : super(key: key);
@@ -78,12 +95,22 @@ class NavigationSide extends StatelessWidget {
                                 e.icon,
                                 color: const Color(0xff8e8e8e),
                               ),
-                              badgeContent: const Text(
-                                '23',
-                                style: TextStyle(
-                                  fontSize: 8,
-                                  color: Colors.white,
+                              badgeBuilder: (tile) => badges.Badge(
+                                badgeContent: const Center(
+                                  child: Text(
+                                    '23',
+                                    style: TextStyle(
+                                      fontSize: 8,
+                                      color: Colors.white,
+                                    ),
+                                  ),
                                 ),
+                                position: badges.BadgePosition.custom(
+                                  end: 12.0,
+                                  bottom: 0.0,
+                                  top: 0.0,
+                                ),
+                                child: tile,
                               ).showOrNull(data.isOpen),
                             ),
                           )