+++ lib/src/component/resizer.dart
@@ -4,11 +4,10 @@ import 'package:flutter_side_menu/src/utils/constants.dart';
 
 class Resizer extends StatefulWidget {
   const Resizer({
-    Key? key,
+    super.key,
     required this.onPanUpdate,
     ResizerData? data,
-  })  : data = data ?? const ResizerData(),
-        super(key: key);
+  })  : data = data ?? const ResizerData();
 
   final Function(DragUpdateDetails details) onPanUpdate;
   final ResizerData data;
@@ -48,7 +47,7 @@ class _ResizerState extends State<Resizer> {
                 : widget.data.resizerColor,
             duration: Constants.duration,
             width: widget.data.resizerWidth,
-            height: MediaQuery.of(context).size.height,
+            height: MediaQuery.sizeOf(context).height,
           ),
         ),
       ),