+++ lib/src/side_menu_controller.dart
@@ -2,4 +2,5 @@ class SideMenuController {
   late void Function() open;
   late void Function() close;
   late void Function() toggle;
+  late bool Function() isCollapsed;
 }